+++ app/src/main/java/com/example/taskill/handlers/DataHandler.java
@@ -7,6 +7,7 @@ import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -24,8 +25,8 @@ public class DataHandler {
         mDatabase.child("users").child(userId).setValue(user);
     }
 
-    public void writeNewServiceProvider(String userId, String name,String username, String email,String password, List<Booking> bookings, Map<String,Double> provided_services) {
-        ServiceProvider user = new ServiceProvider(name,username,email,password,bookings,provided_services);
+    public void writeNewServiceProvider(String userId, String name,String username, String email,String password, List<Booking> bookings, Map<String,Integer> provided_services) {
+        ServiceProvider user = new ServiceProvider(name,username,email,password,bookings, provided_services);
         mDatabase.child("users").child(userId).setValue(user);
     }
